Updates links, downloads and usage pages
This removes most devices from the links pages (as they're not sold
anymore) and replaces them by current devices of 'big brands' mostly.
Furthermore this page is cleaned, dead links and dead projects are
removed. The downloads and usage pages are updated in a similar
fashion, but I took the liberty to reorganize them and add some
applications as well.

Signed-off-by: Erik de Castro Lopo <erikd@mega-nerd.com>