Restore GET method for news feed xhr.
The site was using POST to fetch the rss feed for display on
the landing and news pages to work around aggressive caching
in the apache webserver config.

Since the since was migrated to nginx this method no longer
works. We're not aggressively caching, so GET works fine.
If we do, an exception should simply be made for these
files and/or a server reload should be triggered on update.